updates based on feedback and discussion
