+++ pkg/registry/testdata/invalid-crd.yaml
