+++ pkg/registry/testdata/sample-crd.yaml
